requirements.txt
boto==2.49.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==19.9.0
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.12.1
MarkupSafe==1.1.1
requests==2.22.0
requests2==2.16.0
SQLAlchemy==1.3.7
urllib3==1.21.1
Werkzeug==0.15.5
